expand newly created event templates
A schedule template shows a list of event templates. Each of these has
many fields (description, people to invite, date & time, etc.) and are
shown in a ‘collapsed’ state by default, for a better overview. You
‘expand’ one to edit its fields.

But when you create a new event template, you will need to edit it right
away, so it should be expanded by default, not collapsed. To achive
this, this diff makes the changes below.

Change the ‘collapsed’ field from ‘inner state’ (controlled by the
<EventTemplate/> component and always having the same default value) to
a ‘property’ that the parent component can set.

Keep some ‘extra event data’ alongside the ‘event data model’.
Previously, the only extra data were error messages, so we had an array
of strings. Replace that with an array of ‘extra data’ objects (which
now have 2 fields: ajaxErr: string and collapsed: boolean).

Task https://app.asana.com/0/12169742304238/14842515879498