re-submit a failed scheduling request
The page has several “scheduling requests” and their statuses.
The users clicks “Re-Schedule” on one of them → AJAX start
Prevent any interactions with the page, especially starting to re-submit
another failed request. Because when the above AJAX completes, we'll
navigate away (to the same URL, to re-load it).
→ AJAX complete
If AJAX error: just show the error
If AJAX success: reload the current page